An Analysis of NFL Injuries
The goal I have for this project is to assess is examining the effects that playing on synthetic turf versus natural turf can have on player movements and the factors that may contribute to lower extremity injuries. I will be comparing different factors like part of the body that was injured, number of days out with injury, roster position of injured player, stadium type, temperature, and weather.